% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bullseye}
\alias{bullseye}
\title{An imbalanced dataset with a minor class centered around the origin with a majority class surrounding the center.}
\format{
a data.frame with 1000 rows and 3 columns.
}
\source{
\url{https://gist.github.com/s-kganz/c2534666e369f8e19491bb29d53c619d}
}
\usage{
bullseye
}
\description{
An imbalanced dataset with a minor class centered around the origin with a majority class surrounding the center.
}
\keyword{datasets}
