% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{do.oversample}
\alias{do.oversample}
\title{Oversampling driver function.}
\usage{
do.oversample(data, method, cls.col, cls, m, ...)
}
\arguments{
\item{data}{Data frame to be oversampled.}

\item{method}{Oversampling method to use. See \code{?SCUT} for more information.}

\item{cls.col}{Column in \code{data} containing class membership.}

\item{cls}{Class to be oversampled.}

\item{m}{Desired number of samples in oversampled data frame.}

\item{...}{Other parameters passed to oversampling function.}
}
\value{
data.frame
}
\description{
Oversampling driver function.
}
\examples{
do.oversample(iris, "SMOTE", "Species", "setosa", 70)
}
