% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{do.undersample}
\alias{do.undersample}
\title{Undersampling driver function.}
\usage{
do.undersample(data, method, cls.col, cls, m, ...)
}
\arguments{
\item{data}{data.frame}

\item{method}{character}

\item{cls.col}{character}

\item{cls}{character}

\item{m}{integer}

\item{...}{Other options passed to custom function.}
}
\value{
data.frame
}
\description{
Undersampling driver function.
}
\examples{
do.undersample(iris, "mclust", "Species", "setosa", 30)
}
