% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imbalance}
\alias{imbalance}
\title{An imbalanced dataset with randomly placed normal distributions around the origin.
The nth class has n * 10 observations.}
\format{
a data.frame with 2100 rows and 11 columns
}
\source{
\url{https://gist.github.com/s-kganz/d08473f9492d48ea0e56c3c8a3fe1a74}
}
\usage{
imbalance
}
\description{
An imbalanced dataset with randomly placed normal distributions around the origin.
The nth class has n * 10 observations.
}
\keyword{datasets}
