% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindist.R
\name{mindist.undersample}
\alias{mindist.undersample}
\title{Undersample a dataset by iteratively removing the observation with the lowest total distance to its neighbors.}
\usage{
mindist.undersample(subset, cls, cls.col, m, dist.calc = "euclidean")
}
\arguments{
\item{subset}{Dataset to undersample. Aside from cls.col, must be numeric.}

\item{cls}{Unused, but kept here for compatability with do.undersample().}

\item{cls.col}{Column containing class information.}

\item{m}{Desired number of observations after undersampling.}

\item{dist.calc}{Method for distance calculation. Must be usable with dist().}
}
\value{
An undersampled dataframe.
}
\description{
Undersample a dataset by iteratively removing the observation with the lowest total distance to its neighbors.
}
\examples{
setosa <- iris[iris$Species == "setosa", ]
nrow(setosa)
undersamp <- mindist.undersample(setosa, "setosa", "Species", 50)
nrow(undersamp)
}
