% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oversample.R
\name{oversample.smote}
\alias{oversample.smote}
\title{Oversample a dataset by SMOTE.}
\usage{
oversample.smote(data, cls, cls.col, m)
}
\arguments{
\item{data}{Dataset to be oversampled.}

\item{cls}{Class to be oversampled.}

\item{cls.col}{Column containing class information.}

\item{m}{Desired number of samples in the oversampled data.}
}
\value{
The oversampled dataset.
}
\description{
Oversample a dataset by SMOTE.
}
\examples{
setosa <- iris[iris$Species == "setosa", ]
nrow(setosa)
smoted <- oversample.smote(setosa, "setosa", "Species", 100)
nrow(smoted)
}
