% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undersample.R
\name{resample.random}
\alias{resample.random}
\title{Randomly resample a dataset.}
\usage{
resample.random(data, cls, cls.col, m)
}
\arguments{
\item{data}{Dataframe to be resampled.}

\item{cls}{Class that should be randomly resampled.}

\item{cls.col}{Column containing class information.}

\item{m}{Desired number of samples.}
}
\value{
Resampled dataframe containing only \code{cls}.
}
\description{
This function is used to resample a dataset by randomly removing or duplicating rows. It is usable for both oversampling and undersampling.
}
\examples{
set.seed(1234)
only2 <- undersample.random(wine, 2, "type", 15)
}
