% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{sample.clusters}
\alias{sample.clusters}
\title{Balanced sampled of m individuals from a vector of classification results.}
\usage{
sample.clusters(classif, m)
}
\arguments{
\item{classif}{vector}

\item{m}{integer}
}
\value{
vector
}
\description{
Balanced sampled of m individuals from a vector of classification results.
}
