% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{SCUT.parallel}
\alias{SCUT.parallel}
\title{Parallel implementation of SCUT.}
\usage{
SCUT.parallel(
  form,
  data,
  ncores = doParallel::detectCores() - 2,
  oversample = "SMOTE",
  undersample = "mclust",
  ...
)
}
\arguments{
\item{form}{formula}

\item{data}{data.frame}

\item{ncores}{integer}

\item{oversample}{character}

\item{undersample}{character}

\item{...}{Options passed to custom over/undersampling function.}
}
\value{
data.frame
}
\description{
This function balances multiclass training datasets.
}
\examples{
SCUT(Species ~ ., iris, oversample="SMOTE", undersample="mclust")
SCUT(feed ~ ., chickwts, oversample="SMOTE", undersample="random")
}
