% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undersample.R
\name{undersample.kmeans}
\alias{undersample.kmeans}
\title{Undersample a dataset by kmeans clustering.}
\usage{
undersample.kmeans(data, cls, cls.col, m, k = 5)
}
\arguments{
\item{data}{Dataset to be undersampled.}

\item{cls}{Class to be undersampled.}

\item{cls.col}{Column containing class information.}

\item{m}{Number of samples in undersampled dataset.}

\item{k}{Number of centers in clustering.}
}
\value{
The undersampled dataframe containing only instances of \code{cls}.
}
\description{
Undersample a dataset by kmeans clustering.
}
\examples{
table(iris$Species)
undersamp <- undersample.kmeans(iris, "setosa", "Species", 15)
nrow(undersamp)
}
