% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{validate.dataset}
\alias{validate.dataset}
\title{Validate a dataset for resampling.}
\usage{
validate.dataset(data, cls.col)
}
\arguments{
\item{data}{Dataframe to validate.}

\item{cls.col}{Column with class information.}
}
\value{
NA
}
\description{
This functions checks that the given column is present in the data and that all columns besides the class column are numeric.
}
