% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wine}
\alias{wine}
\title{Type and chemical analysis of three different kinds of wine.}
\format{
a data.frame with 178 rows and 14 columns
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Wine}
}
\usage{
wine
}
\description{
Type and chemical analysis of three different kinds of wine.
}
\keyword{datasets}
