% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{do.oversample}
\alias{do.oversample}
\title{Oversampling driver function.}
\usage{
do.oversample(data, method, cls.col, cls, m, ...)
}
\arguments{
\item{data}{data.frame}

\item{method}{character}

\item{cls.col}{character}

\item{cls}{character}

\item{m}{integer}

\item{...}{}
}
\value{
data.frame
}
\description{
Oversampling driver function.
}
