% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{do.undersample}
\alias{do.undersample}
\title{Undersampling driver function.}
\usage{
do.undersample(data, method, cls.col, cls, m, ...)
}
\arguments{
\item{data}{Data frame to be undersampled.}

\item{method}{Undersampling method to use. See \code{?SCUT} for more information.}

\item{cls.col}{Column in \code{data} containing class membership.}

\item{cls}{Class to undersample.}

\item{m}{Desired number of rows in undersampled data frame.}

\item{...}{Other parameters passed to undersampling function.}
}
\value{
Undersampled data frame.
}
\description{
Undersampling driver function.
}
\examples{
do.undersample(iris, "mclust", "Species", "setosa", 30)
}
