% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{do.undersample}
\alias{do.undersample}
\title{Undersampling driver function.}
\usage{
do.undersample(data, method, cls.col, cls, m, ...)
}
\arguments{
\item{data}{data.frame}

\item{method}{character}

\item{cls.col}{character}

\item{cls}{character}

\item{m}{integer}

\item{...}{}
}
\value{

}
\description{
Undersampling driver function.
}
