% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{SCUT}
\alias{SCUT}
\alias{SCUT.parallel}
\title{SMOTE and cluster-based undersampling technique.}
\usage{
SCUT(
  data,
  cls.col,
  oversample = oversample.smote,
  undersample = undersample.mclust,
  osamp.opts = list(),
  usamp.opts = list()
)

SCUT.parallel(
  data,
  cls.col,
  ncores = 2,
  oversample = oversample.smote,
  undersample = undersample.mclust,
  osamp.opts = list(),
  usamp.opts = list()
)
}
\arguments{
\item{data}{Numeric data frame containing all variables given in \code{form}.}

\item{cls.col}{The column in \code{data} with class membership.}

\item{oversample}{Oversampling method. Must be a function with the signature \verb{foo(data, cls, cls.col, m, ...} that returns a data frame, one of the \verb{oversample.*} functions, or \code{sample.random}.}

\item{undersample}{Undersampling method. Must be a function with the signature \code{foo(data, cls, cls.col, m, ...)} that returns a data frame, one of the \verb{undersample.*} functions, or \code{sample.random}.}

\item{osamp.opts}{Custom options passed to the oversampling function.}

\item{usamp.opts}{Custom options passed to the undersampling function.}

\item{ncores}{Number of cores to use with \code{SCUT.parallel}.}
}
\value{
A dataframe with equal class distribution.
}
\description{
This function balances multiclass training datasets. In a dataframe with \code{n} classes and \code{m} rows, the resulting dataframe will have \code{m / n} rows per class. \code{SCUT.parallel()} distributes each over/undersampling task across multiple cores. Speedup usually occurs only if there are many classes using one of the slower resampling techniques (e.g. \code{mclust}).
}
\examples{
ret <- SCUT(iris, "Species")
ret2 <- SCUT(chickwts, "feed", undersample=undersample.kmeans)
table(ret$Species)
table(ret2$feed)
\donttest{
ret <- SCUT.parallel(wine, "type", undersample=undersample.kmeans)
table(ret$type)
}
}
