% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{SCUT}
\alias{SCUT}
\title{SMOTE and cluster-based undersampling technique.}
\usage{
SCUT(form, data, oversample = "SMOTE", undersample = "mclust")
}
\arguments{
\item{form}{formula}

\item{data}{data.frame}

\item{oversample}{character}

\item{undersample}{character}
}
\value{
data.frame
}
\description{
This function balances multiclass training datasets.
}
\examples{
SCUT(Species ~ ., iris, oversample="SMOTE", undersample="mclust")
SCUT(feed ~ ., chickwts, oversample="SMOTE", undersample="random")
}
