% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scut.R
\name{SCUT}
\alias{SCUT}
\alias{SCUT_parallel}
\title{SMOTE and cluster-based undersampling technique.}
\usage{
SCUT(
  data,
  cls_col,
  oversample = oversample_smote,
  undersample = undersample_mclust,
  osamp_opts = list(),
  usamp_opts = list()
)

SCUT_parallel(
  data,
  cls_col,
  ncores = detectCores()\%/\%2,
  oversample = oversample_smote,
  undersample = undersample_mclust,
  osamp_opts = list(),
  usamp_opts = list()
)
}
\arguments{
\item{data}{Numeric data frame containing all variables given in \code{form}.}

\item{cls_col}{The column in \code{data} with class membership.}

\item{oversample}{Oversampling method. Must be a function with the signature \verb{foo(data, cls, cls.col, m, ...} that returns a data frame, one of the \verb{oversample.*} functions, or \code{sample.random}.}

\item{undersample}{Undersampling method. Must be a function with the signature \code{foo(data, cls, cls_col, m, ...)} that returns a data frame, one of the \verb{undersample.*} functions, or \code{sample.random}.}

\item{osamp_opts}{Custom options passed to the oversampling function.}

\item{usamp_opts}{Custom options passed to the undersampling function.}

\item{ncores}{Number of cores to use with \code{SCUT_parallel}.}
}
\value{
A dataframe with equal class distribution.
}
\description{
This function balances multiclass training datasets. In a dataframe with \code{n} classes and \code{m} rows, the resulting dataframe will have \code{m / n} rows per class. \code{SCUT.parallel()} distributes each over/undersampling task across multiple cores. Speedup usually occurs only if there are many classes using one of the slower resampling techniques (e.g. \code{mclust}).
}
\examples{
ret <- SCUT(iris, "Species")
ret2 <- SCUT(chickwts, "feed", undersample = undersample_kmeans)
table(ret$Species)
table(ret2$feed)
ret <- SCUT_parallel(wine, "type", ncores = 2, undersample = undersample_kmeans)
table(ret$type)
}
