% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undersample.R
\name{undersample.hclust}
\alias{undersample.hclust}
\title{Undersample a dataset by hierarchical clustering.}
\usage{
undersample.hclust(
  data,
  cls,
  cls.col,
  m,
  k = 5,
  h = NA,
  dist.calc = "euclidean"
)
}
\arguments{
\item{data}{Dataset to be undersampled.}

\item{cls}{Majority class that will be undersampled.}

\item{cls.col}{Column in data containing class memberships.}

\item{m}{Desired number of samples in undersampled dataset.}

\item{k}{Desired number of clusters to derive from clustering.}

\item{h}{Desired height at which to cut the clustering tree. k must be NA for this to be used.}

\item{dist.calc}{Distance calculation method.}
}
\value{
Undersampled dataframe containing only cls.
}
\description{
Undersample a dataset by hierarchical clustering.
}
\examples{
table(iris$Species)
undersamp <- undersample.hclust(iris, "setosa", "Species", 15)
nrow(undersamp)
}
